+++ models/tt_dit/models/vae/vae_flux2.py
@@ -9,13 +9,13 @@ from typing import TYPE_CHECKING
 
 import ttnn
 
-from ...blocks.vae import VaeContext, VaeConv2d, VaeMidBlock, VaeNormDescGroup, VaeUpBlock
+from ...blocks.vae_opt import VaeContext, VaeConv2d, VaeMidBlock, VaeNormDescGroup, VaeUpBlock, _all_gather_hw, _norm
 from ...layers.linear import Linear
 from ...layers.module import Module, ModuleList, Parameter
-from ...layers.normalization import GroupNorm
-from ...parallel.config import VAEParallelConfig
+from ...parallel.config import Flux2VaeParallelConfig
 from ...parallel.manager import CCLManager
 from ...utils.substate import pop_substate, rename_substate
+from ...utils.tracing import traced_function
 
 if TYPE_CHECKING:
     from collections.abc import Sequence
@@ -34,16 +34,22 @@ class Flux2VaeDecoder(Module):
         block_out_channels: Sequence[int] = (128, 256, 512, 512),
         layers_per_block: int = 2,
         z_channels: int = 32,
-        parallel_config: VAEParallelConfig | None,
+        parallel_config: Flux2VaeParallelConfig,
         device: ttnn.MeshDevice,
         ccl_manager: CCLManager | None,
+        use_conv3d: bool = False,
     ) -> None:
         super().__init__()
 
         ctx = VaeContext(
-            tp_axis=parallel_config.tensor_parallel.mesh_axis if parallel_config is not None else None,
+            tp_axis=parallel_config.tp_parallel.mesh_axis if parallel_config.tp_parallel is not None else None,
             device=device,
             ccl_manager=ccl_manager,
+            h_mesh_axis=parallel_config.h_parallel.mesh_axis if parallel_config.h_parallel is not None else None,
+            h_factor=parallel_config.h_parallel.factor if parallel_config.h_parallel is not None else 1,
+            w_mesh_axis=parallel_config.w_parallel.mesh_axis if parallel_config.w_parallel is not None else None,
+            w_factor=parallel_config.w_parallel.factor if parallel_config.w_parallel is not None else 1,
+            use_conv3d=use_conv3d,
         )
 
         if ctx.tp_axis is not None and ctx.ccl_manager is None:
@@ -74,9 +80,7 @@ class Flux2VaeDecoder(Module):
             for i, (ch_in, ch_out) in enumerate(itertools.pairwise(channel_counts))
         )
 
-        self.conv_norm_out = GroupNorm(
-            num_groups=32, num_channels=channel_counts[-1], eps=1e-6, mesh_axis=ctx.tp_axis, mesh_device=ctx.device
-        )
+        self.conv_norm_out = _norm(VaeNormDescGroup(num_groups=32, eps=1e-6), num_channels=channel_counts[-1], ctx=ctx)
         self.conv_out = VaeConv2d(
             channel_counts[-1], out_channels, kernel_size=3, padding=1, tensor_parallel=False, ctx=ctx
         )
@@ -86,8 +90,7 @@ class Flux2VaeDecoder(Module):
         self.bn_running_var = Parameter(total_shape=[bn_size], device=ctx.device)
         self.bn_eps = 1e-4
 
-        self._tp_axis = ctx.tp_axis
-        self._ccl_manager = ctx.ccl_manager
+        self._ctx = ctx
 
     def _prepare_torch_state(self, state: dict[str, torch.Tensor]) -> None:
         # remove encoder state
@@ -111,34 +114,45 @@ class Flux2VaeDecoder(Module):
         return z * s + m
 
     def preprocess_and_unpatchify(self, z: ttnn.Tensor, *, height: int, width: int) -> ttnn.Tensor:
-        # N, (H / P) * (W / P), C * P * P -> N, H, W, C
-
-        n, _, _ = z.shape
+        # N, (H / P) * (W / P), C * P * P -> N, H/h_factor, W, C
+        # Input may be H-sharded (h_factor rows per device) or fully replicated.
+        # W-sharding is applied by the caller after this call, once the spatial layout
+        # is restored (H/W are interleaved in patchified tokens so W can't be split here).
         p = self._PATCH_SIZE
+        assert (
+            height % (p * self._ctx.h_factor) == 0
+        ), f"height {height} must be divisible by patch {p} * h_factor {self._ctx.h_factor}"
+        assert (
+            width % (p * self._ctx.w_factor) == 0
+        ), f"width {width} must be divisible by patch {p} * w_factor {self._ctx.w_factor}"
+        height_local = height // self._ctx.h_factor
 
         z = self._inv_normalize(z)
-
         z = ttnn.to_layout(z, ttnn.ROW_MAJOR_LAYOUT)
-        z = z.reshape([n, height // p, width // p, -1, p, p])
+        z = z.reshape([z.shape[0], height_local // p, width // p, -1, p, p])
         z = ttnn.permute(z, [0, 1, 4, 2, 5, 3])
-        z = z.reshape([n, height, width, -1])
+        z = z.reshape([z.shape[0], height_local, width, -1])
         z = ttnn.to_layout(z, ttnn.TILE_LAYOUT)
-
         return z
 
+    @traced_function(device=lambda self: self._ctx.device, clone_prep_inputs=False)
     def forward(self, z: ttnn.Tensor, /) -> ttnn.Tensor:
+        # Input arrives H-sharded when SP is active (preprocess_and_unpatchify uses per-device
+        # H), or replicated when SP is off — both cases work for post_quant_conv (per-position
+        # Linear) and the SP-aware conv pyramid below.
         z = self.post_quant_conv.forward(z)
-        z = self.conv_in.forward(z)
 
+        z = self.conv_in.forward(z)
         z = self.mid_block.forward(z)
 
         for block in self.up_blocks:
             z = block.forward(z)
 
         z = self.conv_norm_out.forward(z)
-        z = ttnn.silu(z)
+        z = ttnn.silu(z, output_tensor=z)
 
-        if self._ccl_manager is not None:
-            z = self._ccl_manager.all_gather(z, dim=-1, mesh_axis=self._tp_axis, use_hyperparams=True)
+        if self._ctx.ccl_manager is not None and self._ctx.tp_axis is not None:
+            z = self._ctx.ccl_manager.all_gather(z, dim=-1, mesh_axis=self._ctx.tp_axis, use_hyperparams=True)
 
-        return self.conv_out.forward(z)
+        z = self.conv_out.forward(z)
+        return _all_gather_hw(self._ctx, z)
